refactoring, add pull_request_template.md, fix mock api
