+++ .github/pull_request_template.md
@@ -1,0 +1,29 @@
+## Description
+
+Please include a summary of the change and which issue is fixed.
+Please also include relevant motivation and context.
+Include link to the task (if there is any).
+
+## Fixes (issue)
+
+Did you fix something? Leave empty if this PR contains just the new feature.
+
+
+## Improvements
+
+Did you improve something? Why did you do this? Why do you think it's better?
+
+## How Has This Been Tested?
+
+Please describe how did you test the code.
+
+## Checklist:
+If you have any comments regarding one of these points just write it down.
+
+- [ ] My code follows the style guidelines of this project
+- [ ] I have performed a self-review of my own code
+- [ ] I have commented my code, particularly in hard-to-understand areas
+- [ ] I have made corresponding changes to the documentation
+- [ ] My changes generate no new warnings
+- [ ] I have added tests that prove my fix is effective or that my feature works
+- [ ] New and existing unit tests pass locally with my changes
