+++ bot/handlers/network_info_handlers.py
@@ -1,4 +1,4 @@
-from collections import defaultdict, Counter
+from collections import Counter
 
 from helpers import *
 from messages import *
